requirements.txt
asgiref==3.2.10
astroid==2.9.3
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.2
django-ckeditor==6.2.0
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-js-asset==2.0.0
django-utils-six==2.0
gunicorn==20.1.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
Pillow==9.0.1
platformdirs==2.5.0
psycopg2==2.9.3
pylint==2.12.2
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typed-ast==1.5.2
typing_extensions==4.0.1
whitenoise==6.0.0
wrapt==1.13.3
